feat: adaptação do rollup terser
