+++ lerna.json
@@ -1,5 +1,5 @@
 {
   "packages": ["packages/*", "shared/*", "demo/*"],
-  "version": "1.1.0-dev.27",
+  "version": "1.1.0-dev.30",
   "npmClients": "npm"
 }
