+++ packages/web-components/package.json
@@ -31,9 +31,11 @@
   },
   "dependencies": {
     "@animaliads/ani-button": "^1.1.0-dev.27",
+    "@animaliads/ani-card": "^1.1.0-dev.27",
     "@animaliads/ani-checkbox": "^1.1.0-dev.27",
     "@animaliads/ani-link": "^1.1.0-dev.27",
     "@animaliads/ani-radio": "^1.1.0-dev.27",
+    "@animaliads/ani-select": "^1.1.0-dev.27",
     "@animaliads/ani-switch": "^1.1.0-dev.27",
     "@animaliads/ani-textarea": "^1.1.0-dev.27",
     "@animaliads/ani-textfield": "^1.1.0-dev.27"
