+++ packages/web-components/rollup.config.js
@@ -1,24 +1,6 @@
-import typescript from 'rollup-plugin-typescript2';
+import { getRollupConfig } from '../../rollup.config';
 import pkg from './package.json';
 
 const input = 'src/index.ts';
 
-const external = [
-  ...Object.keys(pkg.dependencies || {}),
-  ...Object.keys(pkg.peerDependencies || {}),
-];
-
-const plugins = [typescript({ typescript: require('typescript') })];
-
-export default [
-  {
-    input,
-    output: {
-      file: pkg.module,
-      format: 'esm',
-      sourcemap: true,
-    },
-    plugins,
-    external,
-  },
-];
+export default getRollupConfig(input, pkg);
