+++ packages/web-components/src/index.ts
@@ -5,3 +5,5 @@ import '@animaliads/ani-radio';
 import '@animaliads/ani-switch';
 import '@animaliads/ani-textarea';
 import '@animaliads/ani-textfield';
+import '@animaliads/ani-card';
+import '@animaliads/ani-select';
